Impovements for security and docs for api

+++ ping-service/src/main/java/ru/holyway/pingservice/scheduler/TaskSchedulerService.java
@@ -6,12 +6,14 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ScheduledFuture;
 import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
+import org.springframework.http.HttpStatus;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.HttpClientErrorException;
 import ru.holyway.pingservice.data.Task;
 import ru.holyway.pingservice.data.TaskInfo;
 import ru.holyway.pingservice.data.TasksRepository;
@@ -27,7 +29,7 @@ public class TaskSchedulerService {
 
   private final UserRepository userRepository;
 
-  private final Map<String, TaskInfo> tasks = new ConcurrentHashMap<>();
+  private final Map<String, TaskInfo> scheduledTasks = new ConcurrentHashMap<>();
 
   public TaskSchedulerService(TaskScheduler taskScheduler,
       TasksRepository tasksRepository, UserRepository userRepository) {
@@ -41,18 +43,12 @@ public class TaskSchedulerService {
     tasksRepository.findAll().forEach(this::addTaskInternal);
   }
 
-  public List<TaskInfo> getTasks() {
-    return tasks.values().stream().filter(taskInfo -> {
-      User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-      if (user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
-        return true;
-      }
-      UserInfo userInfo = taskInfo.getTask().getUser();
-      if (userInfo == null) {
-        return true;
-      }
-      return user.getUsername().equals(userInfo.getName());
-    }).collect(Collectors.toList());
+  public List<Task> getTasks() {
+
+    List<Task> tasks = scheduledTasks.values().stream().map(TaskInfo::getTask).collect(
+        Collectors.toList());
+
+    return tasks.stream().filter(this::checkAccess).collect(Collectors.toList());
   }
 
   public void addTask(final Task task) {
@@ -66,18 +62,83 @@ public class TaskSchedulerService {
   }
 
   private void addTaskInternal(final Task task) {
-    final ScheduledFuture scheduledFuture = taskScheduler
-        .schedule(task, new CronTrigger(task.getCron()));
+    ScheduledFuture scheduledFuture = null;
+
+    if (task.getIsActive() == null || task.getIsActive()) {
+      scheduledFuture = taskScheduler
+          .schedule(task, new CronTrigger(task.getCron()));
+    }
     tasksRepository.save(task);
-    tasks.put(task.getName(), new TaskInfo(task, scheduledFuture));
+    scheduledTasks.put(task.getName(), new TaskInfo(task, scheduledFuture));
+  }
+
+  public void updateTask(final Task task) {
+    final TaskInfo taskInfo = scheduledTasks.get(task.getName());
+    if (taskInfo != null && checkAccess(taskInfo.getTask())) {
+      final ScheduledFuture scheduledFuture = taskInfo.getScheduledFuture();
+      if (scheduledFuture != null) {
+        scheduledFuture.cancel(false);
+      }
+      addTask(task);
+    } else {
+      throw new HttpClientErrorException(HttpStatus.NOT_FOUND);
+    }
+  }
+
+  public void startTask(final String name) {
+    final TaskInfo taskInfo = scheduledTasks.get(name);
+    if (taskInfo != null && checkAccess(taskInfo.getTask())) {
+      final Task task = taskInfo.getTask();
+      task.setIsActive(true);
+      updateTask(task);
+    } else {
+      throw new HttpClientErrorException(HttpStatus.NOT_FOUND);
+    }
+  }
+
+  public void stopTask(final String name) {
+    final TaskInfo taskInfo = scheduledTasks.get(name);
+    if (taskInfo != null && checkAccess(taskInfo.getTask())) {
+      final Task task = taskInfo.getTask();
+      task.setIsActive(false);
+      updateTask(task);
+    } else {
+      throw new HttpClientErrorException(HttpStatus.NOT_FOUND);
+    }
   }
 
   public void removeTask(final String name) {
-    final TaskInfo taskInfo = tasks.get(name);
-    if (taskInfo != null) {
-      final ScheduledFuture scheduledFuture = tasks.get(name).getScheduledFuture();
-      scheduledFuture.cancel(false);
+    final TaskInfo taskInfo = scheduledTasks.get(name);
+    if (taskInfo != null && checkAccess(taskInfo.getTask())) {
+      final ScheduledFuture scheduledFuture = scheduledTasks.get(name).getScheduledFuture();
+      if (scheduledFuture != null) {
+        scheduledFuture.cancel(false);
+      }
       tasksRepository.delete(name);
+      scheduledTasks.remove(name);
+    } else {
+      throw new HttpClientErrorException(HttpStatus.NOT_FOUND);
     }
+  }
+
+  /**
+   * LOLOLOLO
+   */
+  private boolean checkAccess(final Task task) {
+    final User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    if (isAdmin(user)) {
+      return true;
+    }
+    return task.getUser() != null && task.getUser().getName().equals(user.getUsername());
+  }
+
+  /**
+   * LOLOLOOLOLOLOLOLLOLOLOLOLOLOLOLOL
+   */
+  private boolean isAdmin(User user) {
+    if (user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+      return true;
+    }
+    return false;
   }
 }
